test(providers): close the Tier 4 fall-through hazard
The adversarial second opinion was right about a real hole. Stubbing only the
headless seam left the visible-window path live: if the retry inside
finishEscapeHatch ever failed, auth.go:302 would launch a real browser and
wait fifteen seconds for the cookie store to change — longer than the test's
own watchdog, so cleanup could restore the developer's HOME while the worker
was still running and let it persist the fixture's synthetic cookies for the
reserved documentation host into the real cache. It cannot happen with the
fixture as written, which is exactly one careless edit away from happening.

Two fixes, both test-only. currentOpenURL (cookies.go:132, an existing
injectable hook) is stubbed to fail, turning the whole visible path into an
immediate false return, and asserted never to be reached. The worker now runs
on a cancellable context joined in a cleanup registered late enough to run
FIRST, so it is stopped and drained before HOME and the seams are restored.

Splitting the probe file is the repo 800-line policy, not preference: the
behavioural probes moved to zz_deadlock_probe_tiers_test.go and the structural
guards stayed put.

Verified by sabotage: commitAuthValuesLocked -> commitAuthValues at auth.go:133
fails the test on its watchdog. auth.go checksum unchanged, diff empty. Package
suite with -race: 860 pass. make lint exit 0 with the full linter toolchain on
PATH.
